+++ mvp/src/main/java/com/mandria/android/mvp/rx/callbacks/OnCompleted.java
@@ -1,5 +1,7 @@
 package com.mandria.android.mvp.rx.callbacks;
 
+import android.support.annotation.NonNull;
+
 import rx.functions.Action1;
 
 /**
@@ -7,4 +9,6 @@ import rx.functions.Action1;
  */
 public interface OnCompleted<U> extends Action1<U> {
 
+    @Override
+    void call(@NonNull U u);
 }